NP-2705 Split files
