+++ src/pages/registration/DescriptionPanel.tsx
@@ -11,7 +11,7 @@ import { DescriptionFieldNames } from '../../types/publicationFieldNames';
 import { Registration } from '../../types/registration.types';
 import { DatePickerField } from './description_tab/DatePickerField';
 import { ProjectsField } from './description_tab/projects_field/ProjectsField';
-import { VocabularyField } from './description_tab/VocabularyField';
+import { VocabularyField } from './description_tab/vocabularies/VocabularyField';
 import { BetaFunctionality } from '../../components/BetaFunctionality';
 
 const DateAndLanguageWrapper = styled.div`
